requirements.txt
# Contact Enrichment Pipeline Dependencies

# LangGraph and LLM
langgraph==0.1.0
langchain==0.1.0

# Data Processing
pymongo==4.6.0
gspread==5.10.0
gspread-asyncio==1.6.1
oauth2client==4.1.3

# HTTP Client
httpx==0.25.2

# Environment and Config
python-dotenv==1.0.0

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
